+++ CHANGELOG.md
@@ -21,6 +21,7 @@ with respect to its command line interface and HTTP interface.
 ### Fixed
 * Server: the PostgreSQL storage module successfully deduplicates proposed DB
   records generated from user input now.
+* All: /all-deploy-queues returns correctly, with a somewhat different data format.
 
 ### Added
 * Server: convert remainder of singularity package to generalmsg style logs